requirements-dev.txt
black ~= 19.10b0
django-stubs ~= 1.4
mypy ~= 0.761
